Bryans suggestions
